refactor: extract static-proxy middleware, SSR handler, and config builder from program.ts
- Inline proxyUiAssetRequest at its single call site
- Extract buildRuntimeClientConfig, resolveActiveRuntime, getFallbackGatewayId
  to services/config.ts
- Extract static asset proxy middleware (lines 198-222) to
  middleware/static-proxy.ts as createStaticAssetProxyHandler()
- Extract SSR fallback handler (lines 226-298) to
  routes/ssr.ts as createSsrFallbackHandler()
- Remove unused imports from program.ts
- buildRuntimeClientConfig now takes authAvailable: boolean instead of
  plugins: PluginResult to avoid circular dependency
