+++ host/src/middleware/static-proxy.ts
@@ -1,4 +1,12 @@
+import type { Context } from "hono";
 import { proxy } from "hono/proxy";
+import type { AuthVariables } from "../lib/auth";
+import { STATIC_ASSET_PATTERN } from "../middleware/security";
+import type { RuntimeConfig } from "../services/config";
+import { getTenantRuntimeErrorResponse, resolveRequestRuntime } from "../services/tenant-runtime";
+import { logger } from "../utils/logger";
+
+type HonoEnv = { Variables: AuthVariables };
 
 export async function proxyRequest(
   req: Request,
@@ -82,4 +90,35 @@ export async function proxyStaticAssetRequest(req: Request, targetBase: string):
   response.headers.set("cache-control", "public, max-age=14400, s-maxage=300");
 
   return response;
+}
+
+export function createStaticAssetProxyHandler(config: RuntimeConfig) {
+  return async (c: Context<HonoEnv>, next: () => Promise<void>) => {
+    const { pathname } = new URL(c.req.url);
+
+    if (
+      pathname === "/" ||
+      pathname === "/api" ||
+      pathname.startsWith("/api/") ||
+      pathname === "/health"
+    ) {
+      return next();
+    }
+
+    const lastSegment = pathname.split("/").pop() ?? "";
+    if (!STATIC_ASSET_PATTERN.test(lastSegment)) {
+      return next();
+    }
+
+    try {
+      const runtime = await resolveRequestRuntime(config, c.req.raw, {
+        verification: "stale-while-revalidate",
+      });
+      return await proxyStaticAssetRequest(c.req.raw, runtime.config.ui.url);
+    } catch (error) {
+      const { message, status } = getTenantRuntimeErrorResponse(error);
+      logger.error(`[Proxy Asset] ${c.req.method} ${c.req.path} — ${message}`);
+      return c.text(message, { status: status as 404 | 500 | 502 });
+    }
+  };
 }
